streamlit/requirements.txt
streamlit>=1.32.0
psycopg2-binary
pyjwt
streamlit-aggrid
autogen
ag2
openai
flask
google-api-python-client
google-auth
pandas
matplotlib
pytest 